Change JobExecutionResponseReader to an interface
