+++ spark/src/main/java/org/opensearch/sql/spark/response/JobExecutionResponseReader.java
@@ -5,75 +5,25 @@
 
 package org.opensearch.sql.spark.response;
 
-import static org.opensearch.sql.datasource.model.DataSourceMetadata.DEFAULT_RESULT_INDEX;
-import static org.opensearch.sql.spark.data.constants.SparkConstants.DATA_FIELD;
-import static org.opensearch.sql.spark.data.constants.SparkConstants.JOB_ID_FIELD;
-
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 import org.json.JSONObject;
-import org.opensearch.action.search.SearchRequest;
-import org.opensearch.action.search.SearchResponse;
-import org.opensearch.client.Client;
-import org.opensearch.common.action.ActionFuture;
-import org.opensearch.index.IndexNotFoundException;
-import org.opensearch.index.query.QueryBuilder;
-import org.opensearch.index.query.QueryBuilders;
-import org.opensearch.search.SearchHit;
-import org.opensearch.search.builder.SearchSourceBuilder;
 
-public class JobExecutionResponseReader {
-  private final Client client;
-  private static final Logger LOG = LogManager.getLogger();
+/** Interface for reading job execution result */
+public interface JobExecutionResponseReader {
+  /**
+   * Retrieves the result from the OpenSearch index based on the job ID.
+   *
+   * @param jobId The job ID.
+   * @param resultLocation The location identifier where the result is stored (optional).
+   * @return A JSONObject containing the result data.
+   */
+  JSONObject getResultWithJobId(String jobId, String resultLocation);
 
   /**
-   * JobExecutionResponseReader for spark query.
+   * Retrieves the result from the OpenSearch index based on the query ID.
    *
-   * @param client Opensearch client
+   * @param queryId The query ID.
+   * @param resultLocation The location identifier where the result is stored (optional).
+   * @return A JSONObject containing the result data.
    */
-  public JobExecutionResponseReader(Client client) {
-    this.client = client;
-  }
-
-  public JSONObject getResultFromOpensearchIndex(String jobId, String resultIndex) {
-    return searchInSparkIndex(QueryBuilders.termQuery(JOB_ID_FIELD, jobId), resultIndex);
-  }
-
-  public JSONObject getResultWithQueryId(String queryId, String resultIndex) {
-    return searchInSparkIndex(QueryBuilders.termQuery("queryId", queryId), resultIndex);
-  }
-
-  private JSONObject searchInSparkIndex(QueryBuilder query, String resultIndex) {
-    SearchRequest searchRequest = new SearchRequest();
-    String searchResultIndex = resultIndex == null ? DEFAULT_RESULT_INDEX : resultIndex;
-    searchRequest.indices(searchResultIndex);
-    SearchSourceBuilder searchSourceBuilder = new SearchSourceBuilder();
-    searchSourceBuilder.query(query);
-    searchRequest.source(searchSourceBuilder);
-    ActionFuture<SearchResponse> searchResponseActionFuture;
-    JSONObject data = new JSONObject();
-    try {
-      searchResponseActionFuture = client.search(searchRequest);
-    } catch (IndexNotFoundException e) {
-      // if there is no result index (e.g., EMR-S hasn't created the index yet), we return empty
-      // json
-      LOG.info(resultIndex + " is not created yet.");
-      return data;
-    } catch (Exception e) {
-      throw new RuntimeException(e);
-    }
-    SearchResponse searchResponse = searchResponseActionFuture.actionGet();
-    if (searchResponse.status().getStatus() != 200) {
-      throw new RuntimeException(
-          "Fetching result from "
-              + searchResultIndex
-              + " index failed with status : "
-              + searchResponse.status());
-    } else {
-      for (SearchHit searchHit : searchResponse.getHits().getHits()) {
-        data.put(DATA_FIELD, searchHit.getSourceAsMap());
-      }
-      return data;
-    }
-  }
+  JSONObject getResultWithQueryId(String queryId, String resultLocation);
 }
